correction SCC (version finale)

+++ src/interfaces/ISCC.java
@@ -3,8 +3,8 @@ package interfaces;
 import enums.EDirection;
 
 public interface ISCC {
-    void requeteCabine(double niveauDestination);
-    void requeteEtage(double niveauSource, EDirection direction);
+    void requeteCabine(int niveauDestination);
+    void requeteEtage(int niveauSource, EDirection direction);
     void declencherArretUrgence();
     void stopperArretUrgence();
 }
